Correct typos
Words with 1 letter added, 1 letter removed, or 2 letters transposed.

Usually in comments/descriptions. In a few cases, in the name of a variable
or member function. (If the typo was repeated in references to the variable
or function, I fixed it in both places.)

I don't have the familiarity with all the programming environments to recompile
and doublecheck. I found the typos by running a statistical analysis on the 'rare'
words in the source code.